Rename PackageGraph to AcyclicPackageGraph
Also move rdeps to prevent name conflicts with new graph work.

Split AcyclicPackageGraph from PackageGraphTrait to separate functionality.
