+++ components/builder-core/src/acyclic_rdeps.rs
