+++ components/builder-core/src/lib.rs
@@ -31,6 +31,8 @@ use habitat_builder_protocol as protocol;
 use habitat_core as hab_core;
 
 pub mod access_token;
+pub mod acyclic_package_graph;
+pub mod acyclic_rdeps;
 pub mod api_client;
 pub mod build_config;
 pub mod error;
@@ -40,9 +42,8 @@ pub mod job;
 pub mod keys;
 pub mod logger;
 pub mod metrics;
-pub mod package_graph;
+pub mod package_graph_trait;
 pub mod privilege;
-pub mod rdeps;
 pub mod rpc;
 pub mod socket;
 pub mod target_graph;
